1011100: Fix orphaned virt bonus pools in hosted.
Bad patch went in assuming that when a subscription disappeared, we
could skip cleanup of a derived pool as the cleanup for the master pool
would revoke those entitlements, and thus clean up the derived pool.

However for older style virt bonus pools in hosted, there is no source
entitlement or stack to do this, they need to be cleaned up normally.