Add a type indicator to Pool.
When debugging we're currently having to figure out what type of pool
we're looking at based on implied attribute rules. Add a db transient
field to the Pool which will help us clarify what type of pool we're
looking at. Third party callers can use this to adjust behaviour in
displaying pools as necessary.

Also added a wrapper to see if a pool is stacked, and get it's stack ID,
as this is a fairly common check as well.

Ultimately it would be nice to have this saved in the db, but this
requires a more complicated rollout as there is a progrematic upgrade
involved. For now, it will just be transient.