Log a unique request/job ID with every statement.
Uses log4j's thread local MDC to store the value, which can be
referenced in the ConversionPattern in log4j.properties. If the value
has not been set, an empty string will be logged.

For HTTP requests we generate a UUID in the logging servlet filter.

For Pinsertter jobs, we use the already unique job name which matches
what you would see in cp_jobs as the job ID.

This allows us to grep for the statements specific to a particular
request or job execution very easily.