Log the org, if we know it.
Unfortunately this can't really be done in the LoggingFilter as we don't
know how to look up entities there. This means the initial Request log
entry will not know what org the request is specific to. However, once
we get into SecurityInterceptor and load the objects being verified, we
are now able to grab the org off each and store it for logging.

Some requests will not feature an org, as they are not specific to an
org.