Add query filters for owner permission.
This helps to prevent accidental access to owners a user should not be
able to see.

Behaviour changes slightly for this in situations where you're trying to
get access to something you can't have. A find by primary key will still
return a forbidden, but a find by other properties will now just return
a not found. This is a small net gain in that it hides information about
what exists and what doesn't from the caller. Given this is only for
situations where we're refusing access, it should not affect any
legitemate users or integrating apps.