1030644: Emergency performance fix for distributors with lots of ents.
Problems with data serialization in the rules have exposed issue with
consumers with lots of entitlements. Current issue is for distributors
with 200+. As we do not care about status for these distributors, this
patch disables compliance checking and just reports that everything is
ok.

This will possibly be reverted, as we need to investigate the actual
issue for other types of consumers as well.