select a more proper pool during testing.
We were getting random 403 errors from spec tests on the jenkins
build box. Debugging revealed rules violations because we randomly
pulled the first one which isn't always a valid one. Since we setup
the subscriptions in candlepin_scenarios, product2 is a more proper
target for binding. So we now look for the subscription that maps
to that product and its pools/entitlements to consume.