Rules uses listByConsumerAndDate to get consumer entitlements
This will require slightly less json serialization.  We could
theoretically remove the check from inside of compliance,
but I like that one as well to block problems from cp
version differences.

Fixes a problem that would have come up with the new global
attribute, I was checking all of them, not only the ones with
a valid date in ComplianceRules.isEntitlementCompliant