Add serialVersionUID to Principal class.
Encountered an error deploying changes to Principal class against a
previous database. Job status serializes and stores the Principal class
leading to a problem with the serialVersionUID not matching, as we did
not define one (like we should have).

The only way I could see around this is to set it to the version it was
being generated as previously. I believe this can vary depending on the
compiler implementation, however this has been tested on both Fedora and
RHEL and appears to work ok after deploying on an older db.