Add ability for trusted auth to still lookup user permissions.
If the cp-lookup-permissions header is passed in addition to cp-user,
the TrustedAuth implementation will still query the user service for a
user object.

The user will carry specific permissions which will appear on the
UserPrincipal created as a result.

If the header is not specified we perform trusted auth as before.