Allow User objects to carry non-persisted permissions.
Previously these were tied to Roles, which are a DB construct. User
services however return User objects, and would have to fake Roles with
PermissionBlueprints just to end up with Permissions on a Principal.

Instead allow UserServiceAdapter implementations to stamp permissions
directly onto a User when returning it, which will be transferred
straight to the Principal once created.

If roles are present, it will create and return Permission objects for
these as well.