findbugs: use more efficient iterator
WMI: org.candlepin.jackson.DynamicFilterData.addFilters(Object, Set) makes
inefficient use of keySet iterator instead of entrySet iterator

This method accesses the value of a Map entry, using a key that was retrieved
from a keySet iterator. It is more efficient to use an iterator on the entrySet
of the map, to avoid the Map.get(key) lookup.