findbugs: remove unused variable
RCN: Redundant nullcheck of exportStream which is known to be null in
org.candlepin.sync.Importer.loadExport(Owner, File, ConflictOverrides)

This method contains a redundant check of a known null value against the
constant null.

We checked a value for !null when it was never assigned to anything
except null. It effectively was never used, so I removed it.