findbugs: gave each class a private static Logger
findbugs suggested that if it was going to be a protected/public static Logger,
then it should also be final. Well that goes down a checkstyle path of requiring
the logger to be all caps: LOG. So either make all of the dependent classes
use LOG

*OR*

I could simply make all of the classes use the same pattern for loggers we have
had so far with all other classes.