findbugs: exposes array
MS: public static org.candlepin.util.ManagerCapability.getCapabilityList() may
expose internal representation by returning ManagerCapability.capabilities

A public static method returns a reference to an array that is part of the
static state of the class. Any code that calls this method can freely modify the
underlying array. One fix is to return a copy of the array.

The fix was simpler, we simply removed the entire ManagerCapability class as
it is just a wrapper for a static array of strings only used by Status class.
Just put the list in Status and move on.