Remove some old code for cleaning up duplicate attributes.
This does not appear to be a problem any more given that derived product
attributes are working fine. I have confirmed no duplicate attributes
appear after a test data deploy and pool refresh. Likely fixed by
changes to the annotations in the years since this was written.

The cleanup code was causing problems on Hibernate 3 (fixed on 4) with
some incoming changes for how we look up entities.