include log4j-bridge so logdriver works again
logdriver uses log4j to log its messages. With the move to
slf4j we no longer saw logdriver debug logging in any of the
log files. To avoid adding log4j and a log4j.properties just
to use logdriver in development, we decided the bridge for
log4j is a better solution.