findbugs: exclude EI_EXPOSE_REP(2) errors
Many of our classes return a Date which are immutable. This means
that a user of a class could call getExpired(), and change the
value of the returned Date affecting the class. In other words,
given a class Foo with an expired data member.

public class Foo {
  private Date expired = new Date("2013-12-19");
  public Date getExpired() {
    return expired;
  }
}

A class Bar could modify the value as follows:

public class Bar {
  ...
  Foo f = new Foo();
  Date expired =f.getExpired();
  expired.setMonth(0);

  System.out.println(f.getExpired());

This would print 2013-01-19. This isn't horrible in our case
since we don't usually do this. The proper way to fix it is to
change ALL the methods that returned a reference to the Date
objects to construct new ones:

  public Date getExpired() {
    return new Date(expired);
  }

We're not going to fix it, so I wanted to quiet down findBugs
from reporting it any more.