findbugs: WMI_WRONG_MAP_ITERATOR
WMI:
org.candlepin.util.X509V3ExtensionUtil.orderStrings(X509V3ExtensionUtil$PathNode)
makes inefficient use of keySet iterator instead of entrySet iterator

This method accesses the value of a Map entry, using a key that was retrieved
from a keySet iterator. It is more efficient to use an iterator on the entrySet
of the map, to avoid the Map.get(key) lookup.