Revert "Remove qpid/amqp code"
This reverts commit 37db29973c8c9931a665fc9fd88324ed106bbe6a.

Conflicts:
	buildfile
	src/main/java/org/candlepin/config/ConfigProperties.java
	src/main/java/org/candlepin/guice/CandlepinContextListener.java
	src/main/java/org/candlepin/guice/CandlepinModule.java

Updated to work with the new jackson. There is an outstanding FIXME
with the AMQPBusEventAdapter.java which needs the upstream ca cert.
Before we only had the one, we how have a directory. Need to decide
if this is still needed or if we need to send the entire list.