1064968: Allow consumers to check in hypervisors
Permission changes blocked consumers from
looking up hypervisor consumers in order to
decide whether to create a new one, or update
guestIds.

Add error message when host is skipped
This will get passed to virt-who so that there is some
context as to why a host was not updated. Also added
a warning log message.

Drop SecurityHole on hypervisor resource.
Add some logging in hypervisor resource.