Make createModifiesDateFilteringCriteria more readable
Use the "between" criteria restriction rather than implementing
our own.

Between can be inclusive or exclusive depending on the database
implementation.   We are safe in either case because the check
for complete overlap, where the entitlement is completely
encapsulated uses <= and >=.

The only logical change is that depending on database impl,
we will no longer check if an entitlement modifies another
entitlement that starts on the first ents end datetime.