Fix autobind host during guest autobind for derived pools.
Derived provided products were not being considered at all during a guest
autobind where we attempt to autobind the host to things that will unlock
virt_only pools for the guest.

Fixed so autobind is now aware of this type of autobind and knows to consider
the derived product info instead of the top level in this case.

Includes some additional comments for autobind javascript from Carter.