Optimize storing, reading, and getting the rules
Removed some strange looking code around storing rules in the
database, particularly around storing up to one at a time.

Now we always put the rules in the database, and getRules
looks up the row with the most recent updated date. Also
avoid getting the write lock for rules unless it actually
might be necessary by making one check outside the lock,
and only getting the lock for a second check if the first
shows a change.