search consumers by on facts
This behavior mirrors that of pool lookup by attributes.
supplied keys are ANDed together, if there are multiple values
for a key, they are considered ORs.  The key is case sensitive,
however values are not.

Fact lookup supports the wildcard *.  It may be used any number
of times in the key or value.