remove ant-nodeps from fedora builds.
We added ant-nodeps so that we could get the Symlink.class from ant.
In Fedora 20, ant-nodeps is no longer available, and Symlink was moved into ant
proper. On Fedora 18 & 19, it is also in ant proper, though the ant package had
a virtual provides for ant-nodeps. But since we already BuildRequires ant,
BuildRequires for ant-nodeps is redundant.

ant-nodeps is still require for RHEL brew builds so we need to make it
conditional.