1092678: Handle failure to delete expired revoked serials.
We have not been able to identify how this is possible, but in meantime we will
gracefully handle the error. With this patch the error will be logged, but will
not block CRL regeneration entirely.

The bad serials remain in the db, but should not be present in the CRL itself.