Fix async bind for "my" system admins.
A use case that would only surface in hosted, if a principal carrying the my
systems permission runs a job, of which there is probably only one (async
bind), our code attempts to serialize the principal to the db. Unfortunately
the new permission did not implement serializable.

Attempted a unit test but this seems to occur pretty deep in the quartz code
where our unit tests typically mock everything out. Switched to a spec test
which made exposing the issue easy.