Fix client registration for my system admins.
In an environment where my system admins are live, using such an account with
the subscription-manager client will need to first check how many owners the
account can register too. The old API call for this was asking for access level
ALL, i.e. org admins, which these accounts do not have. The result is failed
registrations for any my systems admin.

Because we must maintain compatability with old clients, we'll stick with
original meaning of this API call which is "what orgs can I register to", but
asking specifically for the ability to create a consumer. Tests included verify
this still returns correct owner for an org admin.