Wait for async bind spec test to finish.
Looks like slower systems may fail on this trying to cleanup owner at end of
test, when the async bind is still underway. This causes an entitlement to be
committed that the delete owner job didn't get loaded in time, and a failed
cleanup due to foreign key constraint vilolations.

Noticed we have two "wait for job" implementations, switched to the better
looking one.