Add an Ant-based Checkstyle plugin.
The Checkstyle plugin provided by Buildr runs Checkstyle via the
command-line and is limited in how configurable it is.

This plugin runs Checkstyle via ant and takes its configuration from
the .checkstyle file (which is what the Eclipse Checkstyle plugin uses).
By using this plugin we are able to store all Checkstyle information
within .checkstyle and keep what Buildr is running in-line with what
Eclipse is running.

Additionally this plugin adds the ability to create Checkstyle reports
in XML or HTML format.