1113202: Add derived content to distributors ent cert
For VDC subscriptions, the host entitlements have no
content. This is what's getting exported to a distributor
and stored as the upstream cert. When a guest eventually
goes to use an entitlement from it's derived VDC pool,
the upstream cert has no content on it, preventing the
guest from getting content it is entitled to.

This patch generates VDC certificates differently if being granted
to a distributor, and includes all content for both the parent and
the derived pools. This will be the upstream certificate that
gets used to access content.

Conflicts:
	src/main/java/org/candlepin/service/impl/DefaultEntitlementCertServiceAdapter.java