Allow UsernameConsumersPermission to list owner info
A SubResource.CONSUMERS subresource was added to the
verify annotation of OwnerInfo.getOwnerInfo to properly
allow filtering of OwnerInfo based on consumer permissions.

Without the subresource, the permission would not allow
getting owner info w/o an OwnerPermission (allowing access
to all consumers, not only consumers created by the user).