Simplify some OwnerInfoCurator methods
We were manually checking pool (and pool product) attributes
in sql when we have a PoolFilterBuilder that simplifies this
work for us.

Added wildcard support to PoolFilterBuilder (allows wildcards
in the api as well)

Fixed a bug in PoolFilterBuilder where we would match a
key/value in Product Attributess that has been overridden by
a pool attribute.