Build .checkstyle files programmatically.
Eclipse projects are created in the subproject directory and are unaware
of any directories above them.  This creates an issue for Eclipse
Checkstyle when we want to reference configuration that is stored at the
top level of the entire source tree (such as checks.xml which is off the
project_conf directory).

One solution would be to create a symlink in every project that points
to the shared configuration directory.  This approach creates a lot of
redundancy and potential for people to forget to create the symlink
directory, etc.

The other approach is to generate the .checkstyle file programmatically
for each project when the Eclipse task is run.  The master .checkstyle
file uses an HTML entity to point to the project_conf directory and
we change the value of that entity when we generate the .checkstyle
files for the subprojects.