Refactor RSpec test task.
We can now run specific tests by appending items to the task name.  For
example, `buildr rspec:vcpu,consumer` will run any spec file that begins
with "vcpu" or "consumer".  We can exclude tests with a minus sign in
front of the identifier.  E.g. `buildr rspec:-vcpu` will run all spec
files that do not begin with "vcpu".

Additionally, we can provide either line numbers or test names to the
task.  For example, `buildr rspec:vcpu:62,41` will run the tests on line
62 and 41 of the vcpu spec file.  Likewise, `buildr
rspec:vcpu,consumer:consumer` will run all tests in the vcpu and
consumer specs that have the word "consumer" in the test name.

The general syntax is

rspec:test_name[,test_name ...][:signifier[,signifier ...]]

where the signifier is either a string or an integer.

Please note that if you need to use a phrase to single out a test, you
will need to quote the task name: `buildr "rspec:vcpu:should be valid"`.
Also note that any phrase or line number you specify will be applied to
*all* tests.  So `buildr "rspec:vcpu,consumer:62" will only run tests
that begin on line 62 in either the vcpu or consumer specs.  This is a
limitation of rspec itself.