Fix consumer modified events
We were sending the wrong 'updated' consumer record.  We dont
want to send an event with just the consumer update (with null
values for every field that isn't being updated).  Now we send
the resulting consumer after the update is applied.

I also added a ConsumerEventBuilder helper that we could
eventually build out to support other types of events.