Fix apicrawl task.
Buildr always runs in the top level of the source tree.  The ApiCrawler
task was previously always trying to write out the JSON to the `target`
directory but since we have reorganized the code, the top level no
longer has a `target` directory and the ApiCrawler would fail when
trying to open the output file.