Reorganize all imports according to the Candlepin import order.
The import order is set in project_conf/eclipse/candlepin.importorder.
Also note that this commit consolidates all static imports into a
wildcard.  I did this because it is very annoying when working with
Junit or Mockito to have to constantly add an import for each
static method.  By default Eclipse will not do wildcards and if you
reorganize the imports it will remove it.  You can change this behavior
by going to Preferences -> Java -> Code Style -> Organize Imports and
set "Number of static imports needed for .*" to 1.