Wrap all REST API requests in a transaction.
Early intention of the API was that a request would either succeed or fail, and
not partially make it to the database. This has either been lost over time or
was never working in the first place.

Guice's @Transactional annotation will no-op if a transaction is already
opened, thus the best way to get this behaviour is to ensure all REST api
methods are wrapped in @Transactional giving us one transaction per request,
and no way for a request to partially succeed.