Fixed max result size issue in report
Fixed an issue with the Consumer Status Report where an error
would be thrown if the result set reached a certain size.

First query for the IDs of the latest compliance snapshots
and then do a final lookup by id using find. The find method
does not have a data size limit here since it returns a DB
cursor to the results.

While doing this is not that pretty, it did not seem to have
much of an impact on performance as a lookup by a set of IDs
is pretty quick.

Indexes were also added to the consumer and compliance collections,
however, this did not seem to improve query performance all that
much.