Fix job factory unit test.
Removing the class comparison in the test, it looks of very little significance
and this would be at least the second time it's needed to be updated. Too
fragile for a complex area of the code where we need to wrap/decorate, we
shouldn't care what specific class is returned.