Only send Compliance event when compliance changes
Create an MD5 hash of the key parts of compliance
and the consumer when compliance is calculated
and check it against the current hash value that
is stored on the consumer before sending compliance
events.

There is a possibility of collisions with this
approach, meaning an extra event will be sent,
but this is not a big concern.