Allow AMQP bus publisher to re-try failed delivery to bus.
Catching all exceptions here results in the message being acknowledged in
ListenerWrapper, despite not actually delivering the message as we intended.

Instead we allow this method to throw exceptions if something went wrong
delivering to the message bus. The message is then not acknowledged and remains
in the hornetq queue and will be retried.