Major fixes for reliable hornetq message delivery.
Enable durable hornetq queues to store the journal on disk. Previously we were
only in memory so any events that were pending (usually failed and were waiting
to be retried) would be lost.

Set the message ack batch size to 0 to have hornetq immediately acknowledge
every message with the server. Waiting for a batch limit to be reached can
cause duplicate messages if the server goes down and there were acks waiting to
be sent, resulting in those messages being retried on next restart even though
they went out successfully.