Store status on event in gutterball.
Events initially are set with status RECEIVED. If it never makes it past this
state, that implies there was some kind of an error, and theoretically the
processing should be retried periodically in the future.

SKIPPED indicates we didn't have a handler for the event target, or we did but
it didn't implement a method for the particular event type for that target.
These too might theretically be re-examined in the future after an application
upgrade if we decided to start processing on some new kind of event.

RECEIVED indicates we did have a handler for the target/type and it was
processed successfully.