Move to two phase event processing in gutterball.
In first phase we try to do only the very basics, read the event JSON, and get
it into our db in it's raw(ish) form, and commit the transaction. Exceptions
should bubble up causing the event to remain in the bus queue.

In second phase we try to do our actual event processing, if applicable.
Exceptions should never bubble up, instead we log them and leave the event in the
db with a state indicating there was an error.