Do not check Entitlement.updated when generating hash
Because we make changes to Entitlements that do not
affect compliance, we can't just check the last
updated date. Instead check the id and the quantity.
Most other data comes from the pool.

It is Ok to check last updated date on the pool
as we want to capture compliance any time that
pool data changes.