1156576: Fix NPE on hypervisor checkin
The NPE was occurring if a request was made with an empty body.

This was fixed by checking for a null host to guest mapping
and throwing a BadRequestException. I had considered
equating the null data as an empty mapping and allowing
the request to succeed, but IMO the request should be
made with at least an empty mapping.