Fix immediate regen of entitlements that modify their own products.
Products can have content which modifies other product IDs also on a particular
SKU. This was resulting in an entitlement being granted, then immediately being
flagged for regeneration, and generated again on the next GET of serials or
entitlements. In the case of a client bind, this resulted in twice the
computation time for generating the cert, once for the bind, again when we GET
serials.

Modify the query used to exclude the entitlement itself.