Improved bind performance for large SKUs with modifier entitlements.
A SKU with large numbers of provided products and content, where a substantial
number of content sets modify other products resulted in excessive querying of
the database looking for Entitlements which provide the product IDs the content
modifies, indicating we should include that content in the cert.

However we don't really need the entitlements here, we simply want to know if
the consumer has an entitlement granting access to a product ID which overlaps
somewhere with the date range of the entitlement being granted.

Code changed to compute the full list of all entitled product IDs for the date
range only once and store in a set in memory. This is then used for all content
in the cert which might need to check for modified product IDs.

For ~1500 content sets, roughly 180 of which modify other products (sometimes
multiple), this cuts bind time from 25s to a much more manageable 4s.

Conflicts:
	server/src/test/java/org/candlepin/service/impl/test/DefaultEntitlementCertServiceAdapterTest.java