Batch lookups done during hypervisor checkins.
When virt-who checks in we were doing one query per hypervisor to see if it has
a consumer already registered, then two queries per each guest ID on that
hypervisor to lookup it's guest consumer, and it's most recent host consumer.

Instead we lookup all this data once for every hypervisor ID and guest ID in
the checkin. The data is then passed around via a new DTO and used instead of
going straight to the db for every entity.

Conflicts:
	server/src/main/java/org/candlepin/resource/ConsumerResource.java
	server/src/main/java/org/candlepin/resource/HypervisorResource.java
	server/src/test/java/org/candlepin/resource/HypervisorResourceTest.java
	server/src/test/java/org/candlepin/resource/test/ConsumerResourceUpdateTest.java
	server/src/test/java/org/candlepin/resource/test/GuestIdResourceTest.java