Fix SSL hostname verification error after upgrade to Qpid 0.30.
Prior to Qpid 0.3, SSL connections to the Qpid daemon were not
performing hostname validation
(https://issues.apache.org/jira/browse/QPID-5960).

The Qpid configuration script was generating a CA certificate and using
it as the server certificate as well, but the CN on the CA was "ca"
instead of a matching hostname, "localhost".