Make Checkstyle okay with static final variables named "log".
Checkstyle considers variables declared as static and final to be
constants and normally requires those variable names to be in all
capital letters.  This patch makes a special exception for static final
variables named "log" as we prefer to use lower case for the Logger
object, and it's nice to declare the logger as final so the JVM can
optimize things a bit.