Add a job to cleanup old consumer check-ins.
We only need the most recent one, so periodically we'll purge anything but the
most recent for each consumer.

Doing this as a delete, in a manner that worked across databases, does not seem
possible, after 4 attempts we gave up. Instead we'll query for the rows to delete
in a batch, then delete them. The batch size should prevent problems with both
memory and the maximum limitations for an IN clause on each DBMS. We repeat
until there are none left to delete.