Allow key-value filters to be negated.
Clients can now negate a filter by using "!" as the first character in
the value portion of the parameter.

For example:
/owners/foo/pools?product=foo&attribute=some_attribute:!true

This feature should be useful since in many cases an undefined
attribute effectively acts as if the attribute were declared "false".
We can handle both conditions by filtering on "!true".