Detect product changes during refresh again.
Since we no longer copy data onto pools during refresh, we need to change the
process quite radically to still detect these changes and flag entitlements for
regen. Instead of comparing a Product object to pool.product* and
pool.derivedProduct*, we're now comparing a detatched Product object coming
from the new in-memory subscription adapter to the Product object in the
database.

To prevent comparing the same product over and over (if you have multiple
subscriptions for it), or from accidentally reading the wrong data, product
comparison is a separate step. The changed products are then passed around to
(too) many refresh methods so each can verify if their products have changed,
and flag for regen if so. In the future it might be best to pass an object that
more accurately represents what changed.

Actual pool refresh becomes a little simpler as it no longer needs to do
attribute, or product name comparisons, just see if the subscription product
and provided products are still the same, or if they are in the changed set.

Tests not yet passing but coming soon.