Eliminate the use of subscriptions in bind calls.
This was caused by a performance optimization whereby we'd try to lookup
product data off the subscription we'd already queried, instead of looking up
each product as we go. However with product data in our db now we have the info
already available causing a huge pile of references to subscriptions and their
adapter to be no longer needed.